requirements.txt
Flask==0.10.1
Flask-WTF==0.9.4
Jinja2==2.7.2
MarkupSafe==0.18
PyYAML==3.10
WTForms==1.0.5
Werkzeug==0.9.4
bitarray==0.8.1
docopt==0.6.1
itsdangerous==0.23
kazoo==1.3.1
pybloom==1.1
rpyc==3.2.3
thrift==0.9.1
wsgiref==0.1.2
zope.interface==4.0.5
